Push protocols and skills implemented, tests fixed + small refactoring.
