+++ tests/test_cli/test_misc.py
@@ -63,6 +63,7 @@ Commands:
   install       Install the dependencies.
   list          List the installed resources.
   login         Login to Registry account
+  push          Push item to Registry.
   remove        Remove a resource from the agent.
   run           Run the agent.
   scaffold      Scaffold a resource for the agent.
